pyproject.toml
[tool.poetry]
name = "datatrail-cli"
version = "0.1.0"
description = "CHIME/FRB Datatrail CLI"
authors = ["CHIME FRB Project Office"]
license = "MIT"
readme = "README.md"
packages = [{ include = "dtcli" }]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.3"
requests = "^2.29.0"
rich = "^13.3.5"
chime-frb-api = "^3.1.0"
pyaml = "^21.10.1"
cadcutils = "^1.5.1.1"
numpy = "^1.24.3"
cadcdata = "^2.4"
cadctap = "^0.9.11"
mergedeep = "^1.3.4"

[tool.poetry.scripts]
datatrail = "dtcli.cli:cli"

[tool.poetry.group.dev.dependencies]
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
